small fixes related to products

+++ app/graphql/types/products/add_product_attributes_type.rb
@@ -2,7 +2,7 @@ module Types
     module Products
         class AddProductAttributesType < Types::BaseInputObject
             argument :name, String, required: true
-            argument :price, Int, required: true
+            argument :price, Float, required: true
             argument :description, String, required: false
             argument :category_id, ID, required: true
             argument :pizza_border_id, ID, required: false
